[DEV] type 정리
