+++ app.ts
@@ -13,21 +13,12 @@ if (result.error) {
 
 let app = express();
 
-//app.use(cors());
-
 app.use(cors({
     origin: 'http://localhost:4000', // 허용할 출처
     methods: ['GET', 'POST', 'PUT', 'DELETE'], // 허용할 HTTP 메서드
     allowedHeaders: ['Content-Type', 'Authorization'], // 허용할 헤더
     credentials: true
 }));
-
-// app.use((req, res, next) => {
-//     res.header('Access-Control-Allow-Origin', 'http://localhost:4000'); // 클라이언트의  도메인
-//     res.header('Access-Control-Allow-Credentials', 'true');
-//     res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, application/json');
-//     next();
-// })
 
 app.use(bodyParser.json());
 app.use(cookieParser());
