+++ apps/carts/types.ts
@@ -1,0 +1,9 @@
+export interface Cart {
+    cartId: number,
+    bookId: number,
+    title: string,
+    summary: string,
+    img: string,
+    price: number,
+    count: number
+}
