--- apps/order/type.ts
@@ -1,5 +1,0 @@
-export type OrderedBookItem = {
-    book_id: number,
-    count: number,
-    order_id: number
-}
