+++ apps/users/types.ts
@@ -1,0 +1,10 @@
+
+export interface LoginInfo {
+    email: string,
+    password: string
+}
+
+export interface UserInfo {
+    email: string,
+    password: string,
+}
